Fix iOS build-number auto-increment overflow, continue from approved build
PlayerSettings.iOS.buildNumber used int.TryParse, which silently fails
and resets to "1" on a 12-digit timestamp-style build number (Int32
overflows past ~2.1 billion). Switched to long.TryParse so the counter
actually continues instead of quietly resetting.

Also reset the stored build number to 202609110704, the last build
that was approved, so the next export continues that lineage rather
than the disconnected 1/2 sequence the overflow bug produced.

+++ Assets/Scripts/Editor/BuildPipeline.cs
@@ -67,7 +67,7 @@ namespace TubityWAI
             }
             else
             {
-                if (int.TryParse(PlayerSettings.iOS.buildNumber, out int iosBuildNum))
+                if (long.TryParse(PlayerSettings.iOS.buildNumber, out long iosBuildNum))
                 {
                     PlayerSettings.iOS.buildNumber = (iosBuildNum + 1).ToString();
                     Debug.Log($"[BuildPipeline] Auto-incremented iOS buildNumber to: {PlayerSettings.iOS.buildNumber}");
